requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bandit==1.9.3
black==26.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
faiss-cpu==1.13.2
fastapi==0.128.0
filelock==3.20.3
flake8==7.3.0
fsspec==2026.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
isort==7.0.0
Jinja2==3.1.6
joblib==1.5.3
librt==0.7.8
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.1
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
prometheus_client==0.24.1
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pytokens==0.3.0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
starlette==0.50.0
stevedore==5.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
